Separate IRequestManager file
