+++ packages/core/src/storages/request_manager_tandem.ts
@@ -4,13 +4,12 @@ import type { CrawleeLogger } from '../log.js';
 import type { Request, Source } from '../request.js';
 import { serviceLocator } from '../service_locator.js';
 import type { IRequestLoader } from './request_loader.js';
+import type { IRequestManager, RequestsLike } from './request_manager.js';
 import type {
     AddRequestsBatchedOptions,
     AddRequestsBatchedResult,
-    IRequestManager,
     RequestQueueOperationInfo,
     RequestQueueOperationOptions,
-    RequestsLike,
 } from './request_provider.js';
 
 /**
